requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.109.2
h11==0.14.0
httptools==0.6.1
idna==3.6
py-automapper==1.2.3
pydantic==2.6.1
pydantic_core==2.16.2
pyodbc==5.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
